requirements.txt
Pillow
imagehash
numpy
fastapi
uvicorn
memory-profiler
